# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Questionários",
    "version": "1.0.0",
    "description": "Documentação da API de Questionários"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "paths": {
    "/users/register": {
      "post": {
        "tags": ["Usuários"],
        "summary": "Registrar um novo usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": ["name", "email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário cadastrado com sucesso."
          },
          "400": {
            "description": "Erro de validação. Verifique os campos enviados."
          },
          "409": {
            "description": "Usuário já está cadastrado."
          },
          "500": {
            "description": "Erro interno no servidor."
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": ["Usuários"],
        "summary": "Logar com um usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso."
          },
          "401": {
            "description": "Credenciais inválidas."
          },
          "500": {
            "description": "Erro interno no servidor."
          }
        }
      }
    },
    "/users/reset-password": {
      "post": {
        "tags": ["Usuários"],
        "summary": "Redefinir a senha de um usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "newPassword": {
                    "type": "string"
                  }
                },
                "required": ["email", "newPassword"] 
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Senha redefinida com sucesso."
          },
          "404": {
            "description": "Usuário não encontrado."
          },
          "400": {
            "description": "Erro de validação. Verifique os campos enviados."
          },
          "500": {
            "description": "Erro interno no servidor."
          }
        }
      }
    }
  }
}